refactor: :construction: Refactor
Refactored code for refetch
