+++ http/http.js
@@ -1,0 +1,2 @@
+export const base_url = "https://potatotodos.vercel.app/backend/todolistdata";
+export const local_base_url = "http://localhost:3000/backend/todolistdata";
